Refactor database initialization and service registration
