+++ startup/__init__.py
@@ -1,0 +1,1 @@
+"""Application startup utilities."""
