+++ tests/cli_file_processor.py
@@ -17,7 +17,7 @@ project_root = Path(__file__).parent.parent
 sys.path.insert(0, str(project_root))
 
 from config import create_config_manager
-from config.service_registration import register_upload_services
+from services.upload.service_registration import register_upload_services
 from core.service_container import ServiceContainer
 from services.file_processor_service import FileProcessorService
