+++ tests/test_persistence.py
@@ -9,7 +9,7 @@ def test_initialize(monkeypatch):
     pm = PersistenceManager()
     monkeypatch.setattr(
         "services.helpers.database_initializer.initialize_database",
-        lambda db: (db, "helper", "reporter"),
+        lambda *a, **k: (a[0], "helper", "reporter"),
     )
     manager, helper, reporter = pm.initialize(DummyDB())
     assert helper == "helper"
