+++ tests/test_upload_endpoint.py
@@ -32,11 +32,11 @@ class DummyUploadService:
 
 def _create_app(monkeypatch):
     # Provide minimal stubs to avoid heavy imports during module load
-    fake_reg = types.ModuleType("config.service_registration")
+    fake_reg = types.ModuleType("services.upload.service_registration")
     fake_reg.register_upload_services = lambda c: c.register_singleton(
         "uploader", object()
     )
-    monkeypatch.setitem(sys.modules, "config.service_registration", fake_reg)
+    monkeypatch.setitem(sys.modules, "services.upload.service_registration", fake_reg)
 
     cont = ServiceContainer()
     service = DummyUploadService()
@@ -87,11 +87,11 @@ class FailingUploadService:
 
 
 def test_upload_returns_error_on_exception(monkeypatch):
-    fake_reg = types.ModuleType("config.service_registration")
+    fake_reg = types.ModuleType("services.upload.service_registration")
     fake_reg.register_upload_services = lambda c: c.register_singleton(
         "uploader", object()
     )
-    monkeypatch.setitem(sys.modules, "config.service_registration", fake_reg)
+    monkeypatch.setitem(sys.modules, "services.upload.service_registration", fake_reg)
 
     cont = ServiceContainer()
     service = FailingUploadService()
@@ -139,11 +139,11 @@ class DummyFileProcessor:
 
 
 def _create_validator_app(monkeypatch):
-    fake_reg = types.ModuleType("config.service_registration")
+    fake_reg = types.ModuleType("services.upload.service_registration")
     fake_reg.register_upload_services = lambda c: c.register_singleton(
         "uploader", object()
     )
-    monkeypatch.setitem(sys.modules, "config.service_registration", fake_reg)
+    monkeypatch.setitem(sys.modules, "services.upload.service_registration", fake_reg)
 
     cont = ServiceContainer()
     cont.register_singleton("file_processor", DummyFileProcessor())
